--- backend/enable_debug.sh
@@ -1,3 +1,0 @@
-D:\4_Projects\4_Backend\gamejunx\backend\gamejunx_env\Scripts\activate.ps1
-cd app
-$env:FLASK_DEBUG = "1"
