--- backend/models/user_model.py
@@ -1,6 +1,0 @@
-from flask import jsonify
-from app import db, flask_app
-import hashlib
-
-
-
